Build 386 (working in Debug; maybe buildable): refactoring and fixing bugs after refactoring

+++ Core/UI/Editor/EditorPanelElement/TextureAssetsBrowser.cpp
@@ -42,7 +42,7 @@ void TextureAssetsBrowser::Render(IFacadeEngineToUI* pFacade, bool* pOpen)
     }
 
     ImGui::SetNextWindowContentSize(ImVec2(0.0f, layoutOuterPadding_ + numLayoutLine_ * (layoutItemSize_.y + layoutItemSpacing_)));
-    const int textLineHeight = ImGui::GetTextLineHeightWithSpacing();
+    const float textLineHeight = ImGui::GetTextLineHeightWithSpacing();
 
 
     // menu bar
